utils/config_manager.py
import yaml
import os

def load_rules(config_path="config.yaml"):
    if not os.path.exists(config_path):
        raise FileNotFoundError(f"Archivo de configuración no encontrado: {config_path}")
    
    with open(config_path, "r") as f:
        try:
            config = yaml.safe_load(f)
            return config
        except yaml.YAMLError as e:
            raise ValueError(f"Error al cargar YAML: {str(e)}")
